Split out the GameBrowser from MainScreen
