+++ CMakeLists.txt
@@ -1621,6 +1621,8 @@ list(APPEND NativeAppSource
 	UI/IAPScreen.h
 	UI/MainScreen.h
 	UI/MainScreen.cpp
+	UI/GameBrowser.h
+	UI/GameBrowser.cpp
 	UI/MiscScreens.h
 	UI/MiscScreens.cpp
 	UI/MiscViews.h
