+++ UI/UI.vcxproj
@@ -42,6 +42,7 @@
     <ClCompile Include="DisplayLayoutScreen.cpp" />
     <ClCompile Include="DriverManagerScreen.cpp" />
     <ClCompile Include="EmuScreen.cpp" />
+    <ClCompile Include="GameBrowser.cpp" />
     <ClCompile Include="GameInfoCache.cpp" />
     <ClCompile Include="GamepadEmu.cpp" />
     <ClCompile Include="GameScreen.cpp" />
@@ -98,6 +99,7 @@
     <ClInclude Include="DisplayLayoutScreen.h" />
     <ClInclude Include="DriverManagerScreen.h" />
     <ClInclude Include="EmuScreen.h" />
+    <ClInclude Include="GameBrowser.h" />
     <ClInclude Include="GameInfoCache.h" />
     <ClInclude Include="GamepadEmu.h" />
     <ClInclude Include="GameScreen.h" />
